+++ +modules.js
@@ -13,7 +13,7 @@ module.exports = {
     require: 'off'
   },
   rules: {
-    'node/no-unsupported-features/es-syntax': [
+    'n/no-unsupported-features/es-syntax': [
       'error',
       // We could check for `ecmaVersion < 2020` to restrict `dynamicImport`,
       //   but if present, typically compiling, and seems support was available