+++ index.js
@@ -52,7 +52,7 @@ module.exports = {
     {
       files: ['*.ts', '*.tsx', '*.vue'],
       rules: {
-        'node/no-unsupported-features/es-syntax': 'off'
+        'n/no-unsupported-features/es-syntax': 'off'
       }
     },
     {